+++ context-runtime/modules/admin/src/admin_runtime.cc
@@ -60,129 +60,6 @@
 
 namespace clio::run::admin {
 
-// ============================================================================
-// Network-path instrumentation
-// ============================================================================
-//
-// Counters and timers for diagnosing where cross-node throughput is being
-// lost. Each direction is updated by exactly one worker thread (net_send
-// or net_recv per DefaultScheduler::DivideWorkers), so the per-peer
-// arrays don't need atomics — only the periodic-dump snapshot crosses
-// thread boundaries, and that uses relaxed atomics on the timestamp +
-// "owner has-flushed" flag.
-//
-// Look for "[NetStats] ..." lines in the runtime log. Emitted every
-// kNetStatsDumpIntervalSec wall-clock seconds from the Send periodic.
-//
-// Per-peer rows:
-//   peer=N  sin=<count>/<MiB>  sout=<count>/<MiB>  rin=<count>/<MiB>  rout=<count>/<MiB>
-//   send_lbm_ms=<sum>  send_ser_ms=<sum>  recv_des_ms=<sum>
-//
-// Aggregate row:
-//   total: ... (sum across peers)
-//
-// To reset between runs just restart the runtime; counters are process-
-// scoped.
-struct NetPeerStats {
-  // SendIn (outbound cross-node task forwards)
-  uint64_t sin_count = 0;
-  uint64_t sin_bytes = 0;
-  uint64_t sin_lbm_ns = 0;    // time inside lbm_transport->Send for sin
-  uint64_t sin_ser_ns = 0;    // time inside container->SaveTask for sin
-  // SendOut (outbound cross-node task responses)
-  uint64_t sout_count = 0;
-  uint64_t sout_bytes = 0;
-  uint64_t sout_lbm_ns = 0;
-  uint64_t sout_ser_ns = 0;
-  // RecvIn (inbound cross-node task forwards from peer)
-  uint64_t rin_count = 0;
-  uint64_t rin_bytes = 0;
-  uint64_t rin_des_ns = 0;    // time inside deserialize/dispatch
-  // RecvOut (inbound cross-node task responses from peer)
-  uint64_t rout_count = 0;
-  uint64_t rout_bytes = 0;
-  uint64_t rout_des_ns = 0;
-};
-
-static constexpr size_t kMaxNetPeers = 64;  // soft cap; we just modulo
-static NetPeerStats g_net_peer_stats[kMaxNetPeers];
-
-static inline NetPeerStats &NetStats(chi::u64 peer) {
-  return g_net_peer_stats[peer % kMaxNetPeers];
-}
-
-static constexpr double kNetStatsDumpIntervalSec = 1.0;
-static std::chrono::steady_clock::time_point g_net_stats_last_dump =
-    std::chrono::steady_clock::now();
-
-static inline void DumpNetStatsIfDue(chi::u64 self_node_id, size_t num_peers) {
-  auto now = std::chrono::steady_clock::now();
-  double since =
-      std::chrono::duration<double>(now - g_net_stats_last_dump).count();
-  if (since < kNetStatsDumpIntervalSec) return;
-  g_net_stats_last_dump = now;
-
-  // Aggregates across peers
-  uint64_t tot_sin_c = 0, tot_sin_b = 0, tot_sout_c = 0, tot_sout_b = 0;
-  uint64_t tot_rin_c = 0, tot_rin_b = 0, tot_rout_c = 0, tot_rout_b = 0;
-  uint64_t tot_sin_lbm = 0, tot_sout_lbm = 0;
-  uint64_t tot_sin_ser = 0, tot_sout_ser = 0;
-  uint64_t tot_rin_des = 0, tot_rout_des = 0;
-
-  size_t cap = std::min<size_t>(num_peers, kMaxNetPeers);
-  for (size_t p = 0; p < cap; ++p) {
-    if (p == self_node_id) continue;
-    const auto &s = g_net_peer_stats[p];
-    if (s.sin_count == 0 && s.sout_count == 0 && s.rin_count == 0 &&
-        s.rout_count == 0) {
-      continue;
-    }
-    HLOG(kDebug,
-         "[NetStats] self={} peer={} sin={}/{}MiB sout={}/{}MiB "
-         "rin={}/{}MiB rout={}/{}MiB "
-         "sin_lbm_ms={:.1f} sout_lbm_ms={:.1f} "
-         "sin_ser_ms={:.1f} sout_ser_ms={:.1f} "
-         "rin_des_ms={:.1f} rout_des_ms={:.1f}",
-         self_node_id, p, s.sin_count, s.sin_bytes >> 20, s.sout_count,
-         s.sout_bytes >> 20, s.rin_count, s.rin_bytes >> 20, s.rout_count,
-         s.rout_bytes >> 20, s.sin_lbm_ns / 1e6, s.sout_lbm_ns / 1e6,
-         s.sin_ser_ns / 1e6, s.sout_ser_ns / 1e6, s.rin_des_ns / 1e6,
-         s.rout_des_ns / 1e6);
-    tot_sin_c += s.sin_count;
-    tot_sin_b += s.sin_bytes;
-    tot_sout_c += s.sout_count;
-    tot_sout_b += s.sout_bytes;
-    tot_rin_c += s.rin_count;
-    tot_rin_b += s.rin_bytes;
-    tot_rout_c += s.rout_count;
-    tot_rout_b += s.rout_bytes;
-    tot_sin_lbm += s.sin_lbm_ns;
-    tot_sout_lbm += s.sout_lbm_ns;
-    tot_sin_ser += s.sin_ser_ns;
-    tot_sout_ser += s.sout_ser_ns;
-    tot_rin_des += s.rin_des_ns;
-    tot_rout_des += s.rout_des_ns;
-  }
-  HLOG(kDebug,
-       "[NetStats] self={} TOTAL sin={}/{}MiB sout={}/{}MiB "
-       "rin={}/{}MiB rout={}/{}MiB "
-       "sin_lbm_ms={:.1f} sout_lbm_ms={:.1f} "
-       "sin_ser_ms={:.1f} sout_ser_ms={:.1f} "
-       "rin_des_ms={:.1f} rout_des_ms={:.1f}",
-       self_node_id, tot_sin_c, tot_sin_b >> 20, tot_sout_c,
-       tot_sout_b >> 20, tot_rin_c, tot_rin_b >> 20, tot_rout_c,
-       tot_rout_b >> 20, tot_sin_lbm / 1e6, tot_sout_lbm / 1e6,
-       tot_sin_ser / 1e6, tot_sout_ser / 1e6, tot_rin_des / 1e6,
-       tot_rout_des / 1e6);
-}
-
-static inline uint64_t HrtNs() {
-  return std::chrono::duration_cast<std::chrono::nanoseconds>(
-             std::chrono::steady_clock::now().time_since_epoch())
-      .count();
-}
-
-
 // Method implementations for Runtime class
 
 // Virtual method implementations (Init, Run, Del, SaveTask, LoadTask, NewCopy,
@@ -192,20 +69,7 @@ static inline uint64_t HrtNs() {
 // Method implementations
 //===========================================================================
 
-void Runtime::StopRecvThreads() {
-  // Signal dedicated recv threads to exit and join them. The threads cache a
-  // raw pointer to the main transport, so they MUST be joined before
-  // IpcManager::ClearTransports() frees it (see RegisterTransportShutdownHook).
-  recv_shutdown_.store(true, std::memory_order_release);
-  if (peer_recv_thread_.joinable()) peer_recv_thread_.join();
-  if (client_recv_thread_.joinable()) client_recv_thread_.join();
-}
-
-Runtime::~Runtime() {
-  // Backstop: normally the IpcManager transport-shutdown hook already joined
-  // these (while the transport was still alive). Safe to call again.
-  StopRecvThreads();
-}
+Runtime::~Runtime() {}
 
 chi::TaskResume Runtime::Create(ctp::ipc::FullPtr<CreateTask> task,
                                 chi::RunContext &rctx) {
@@ -252,90 +116,12 @@ chi::TaskResume Runtime::Create(ctp::ipc::FullPtr<CreateTask> task,
   // Spawn periodic ClientSend task for client response sending via lightbeam
   client_.AsyncClientSend(chi::PoolQuery::Local(), 100);
 
-  // Dedicated single peer recv thread: polls the main p2p transport
-  // (port 9413 by default) and dispatches inbound task forwards/responses.
-  peer_recv_thread_ = std::thread([this]() {
-    ctp::SystemInfo::SetCurrentThreadName("chi-peer-recv");
-    auto *ipc_manager = CLIO_IPC;
-    ctp::lbm::Transport *lbm_transport = nullptr;
-    for (int spin = 0; spin < 1000 && !recv_shutdown_.load(); ++spin) {
-      lbm_transport = ipc_manager->GetMainTransport();
-      if (lbm_transport) break;
-      std::this_thread::sleep_for(std::chrono::milliseconds(1));
-    }
-    if (!lbm_transport) {
-      HLOG(kError, "[PeerRecvThread] main transport never appeared");
-      return;
-    }
-    HLOG(kInfo, "[PeerRecvThread] started");
-    while (!recv_shutdown_.load(std::memory_order_acquire)) {
-      chi::LoadTaskArchive archive;
-      auto info = lbm_transport->Recv(archive);
-      int rc = info.rc;
-      if (rc == EAGAIN) {
-        std::this_thread::sleep_for(std::chrono::microseconds(1));
-        continue;
-      }
-      if (rc != 0) {
-        if (rc != -1) {
-          HLOG(kError, "[PeerRecvThread] Recv failed rc={}", rc);
-        }
-        continue;
-      }
-      RecvTask dummy_recv;
-      ctp::ipc::FullPtr<RecvTask> recv_fp(&dummy_recv);
-      chi::MsgType msg_type = archive.GetMsgType();
-      switch (msg_type) {
-        case chi::MsgType::kSerializeIn:
-          RecvIn(recv_fp, archive, lbm_transport);
-          break;
-        case chi::MsgType::kSerializeOut:
-          RecvOut(recv_fp, archive, lbm_transport);
-          break;
-        case chi::MsgType::kHeartbeat:
-          break;
-        default:
-          HLOG(kError, "[PeerRecvThread] unknown msg_type={}",
-               static_cast<int>(msg_type));
-          break;
-      }
-      // Release the per-bulk zmq_msg_t handles that ZeroMqTransport::RecvBulks
-      // allocated for any borrowed (caller-didn't-preallocate) bulks. RecvIn /
-      // RecvOut have already copied the bulk payloads into shm-owned task
-      // buffers (via LoadTaskArchive::bulk -> CLIO_IPC->AllocateBuffer +
-      // memcpy + TASK_DATA_OWNER), so dropping the ZMQ frames here is safe.
-      // Without this, every PutBlob/GetBlob loopback round-trip leaks one
-      // 2 MiB ZMQ-owned frame per direction (~8 MiB per op under
-      // direct0+FORCE_NET) — archive goes out of scope after this iteration
-      // anyway, but the desc-owned zmq_msg_t lives in libzmq memory that
-      // only zmq_msg_close releases; the destructor can't reach it.
-      lbm_transport->ClearRecvHandles(archive);
-    }
-    HLOG(kInfo, "[PeerRecvThread] shutting down");
+  // Start dedicated recv threads in IpcManagerRun2Run.
+  CLIO_IPC->GetRun2Run()->StartRecvThreads();
+  // Stop recv threads before the main transport is freed.
+  CLIO_IPC->RegisterTransportShutdownHook([this]() {
+    CLIO_IPC->GetRun2Run()->StopRecvThreads();
   });
-
-  // Dedicated single client recv thread: drains TCP (port 9416) and IPC
-  // (unix socket) client transports via IpcCpu2CpuZmq::RuntimeRecv.
-  client_recv_thread_ = std::thread([this]() {
-    ctp::SystemInfo::SetCurrentThreadName("chi-client-recv");
-    auto *ipc_manager = CLIO_IPC;
-    HLOG(kInfo, "[ClientRecvThread] started");
-    while (!recv_shutdown_.load(std::memory_order_acquire)) {
-      chi::u32 tasks_received = 0;
-      bool did_work = chi::IpcCpu2CpuZmq::RuntimeRecv(ipc_manager,
-                                                       tasks_received);
-      if (!did_work) {
-        std::this_thread::sleep_for(std::chrono::microseconds(1));
-      }
-    }
-    HLOG(kInfo, "[ClientRecvThread] shutting down");
-  });
-
-  // Stop these recv threads at the very start of IpcManager::ClearTransports(),
-  // i.e. before the main transport they poll is freed. Without this, on
-  // shutdown the threads keep calling Recv() on a freed transport -- benign on
-  // Linux but a hard crash on macOS (busy-spin on ENOTSOCK over freed memory).
-  CLIO_IPC->RegisterTransportShutdownHook([this]() { StopRecvThreads(); });
 
   // Spawn periodic WreapDeadIpcs task with 1 second period
   // This task reaps shared memory segments from dead processes
@@ -578,368 +364,17 @@ chi::TaskResume Runtime::Flush(ctp::ipc::FullPtr<FlushTask> task,
 
 /**
  * Helper function: Send task inputs to remote node
- * @param origin_task Task to send to remote nodes
- * @param rctx RunContext for managing subtasks
  */
 void Runtime::SendIn(ctp::ipc::FullPtr<chi::Task> origin_task,
                      chi::RunContext &rctx) {
-  auto *ipc_manager = CLIO_IPC;
-  auto *pool_manager = CLIO_POOL_MANAGER;
-
-  // Validate origin_task
-  if (origin_task.IsNull()) {
-    HLOG(kError, "SendIn: origin_task is null");
-    return;
-  }
-
-  // Get the container associated with the origin_task
-  chi::Container *container =
-      pool_manager->GetStaticContainer(origin_task->pool_id_);
-  if (container == nullptr) {
-    HLOG(kError, "SendIn: container not found for pool_id {}",
-         origin_task->pool_id_);
-    return;
-  }
-
-  // Pre-allocate send_map_key using origin_task pointer
-  // This ensures consistent net_key across all replicas
-  size_t send_map_key = size_t(origin_task.ptr_);
-
-  // Add the origin task to send_map before creating copies.
-  // SendIn runs on net_send_worker; RecvOut/ProcessRetryQueues touch this
-  // map from other threads with the worker split, so the mutex is required.
-  {
-    std::lock_guard<std::mutex> lk(send_map_mutex_);
-    send_map_[send_map_key] = origin_task;
-  }
-
-  // Get pool_queries from task's RunContext
-  if (!origin_task->GetRunCtx()) {
-    HLOG(kError, "SendIn: origin_task has no RunContext");
-    return;
-  }
-  chi::RunContext *origin_task_rctx = origin_task->GetRunCtx();
-
-  const std::vector<chi::PoolQuery> &pool_queries =
-      origin_task_rctx->pool_queries_;
-  size_t num_replicas = pool_queries.size();
-
-  // Reserve space for all replicas in subtasks vector BEFORE the loop
-  // This ensures subtasks_.size() reflects the correct total replica count
-  origin_task_rctx->subtasks_.resize(num_replicas);
-
-  HLOG(kDebug, "[SendIn] Task {} to {} replicas", origin_task->task_id_,
-       num_replicas);
-
-  // Send to each target in pool_queries
-  for (size_t i = 0; i < num_replicas; ++i) {
-    const chi::PoolQuery &query = pool_queries[i];
-
-    // Determine target node_id based on query type
-    chi::u64 target_node_id = 0;
-
-    if (query.IsLocalMode()) {
-      target_node_id = ipc_manager->GetNodeId();
-    } else if (query.IsPhysicalMode()) {
-      target_node_id = query.GetNodeId();
-    } else if (query.IsDirectIdMode()) {
-      chi::ContainerId container_id = query.GetContainerId();
-      target_node_id =
-          pool_manager->GetContainerNodeId(origin_task->pool_id_, container_id);
-    } else if (query.IsRangeMode()) {
-      chi::u32 offset = query.GetRangeOffset();
-      chi::ContainerId container_id(offset);
-      target_node_id =
-          pool_manager->GetContainerNodeId(origin_task->pool_id_, container_id);
-    } else if (query.IsBroadcastMode()) {
-      HLOG(kError,
-           "Admin: Broadcast mode should be handled by "
-           "TaskDispatcher, not SendIn");
-      continue;
-    } else if (query.IsDirectHashMode()) {
-      HLOG(kError,
-           "Admin: DirectHash mode should be handled by "
-           "TaskDispatcher, not SendIn");
-      continue;
-    } else {
-      HLOG(kError, "Admin: Unsupported or unrecognized query type for SendIn");
-      continue;
-    }
-
-    // Get host information for target node
-    const chi::Host *target_host = ipc_manager->GetHost(target_node_id);
-    if (!target_host) {
-      HLOG(kError, "[SendIn] Task {} FAILED: Host not found for node_id {}",
-           origin_task->task_id_, target_node_id);
-      continue;
-    }
-
-    // Create task copy first (needed for both send and retry)
-    ctp::ipc::FullPtr<chi::Task> task_copy =
-        container->NewCopyTask(origin_task->method_, origin_task, true);
-    origin_task_rctx->subtasks_[i] = task_copy;
-
-    // Set net_key in task_id to match send_map_key
-    chi::TaskId &copy_id = task_copy->task_id_;
-    copy_id.net_key_ = send_map_key;
-    copy_id.replica_id_ = i;
-
-    // Update the copy's pool query to current query
-    task_copy->pool_query_ = query;
-
-    // Set return node ID in the pool query
-    chi::u64 this_node_id = ipc_manager->GetNodeId();
-    task_copy->pool_query_.SetReturnNode(this_node_id);
-
-    // Check aliveness before sending
-    if (!ipc_manager->IsAlive(target_node_id)) {
-      float net_timeout = origin_task->pool_query_.GetNetTimeout();
-      if (net_timeout >= 0 && net_timeout < 0.001f) {
-        // net_timeout=0: fail immediately for dead nodes, skip this replica
-        HLOG(kWarning,
-             "[SendIn] Task {} target node {} is dead, net_timeout=0 -> skip",
-             origin_task->task_id_, target_node_id);
-        origin_task_rctx->completed_replicas_++;
-        continue;
-      }
-      HLOG(kWarning,
-           "[SendIn] Task {} target node {} is dead, queuing for retry",
-           origin_task->task_id_, target_node_id);
-      {
-        std::lock_guard<std::mutex> _rqlk(retry_queues_mutex_);
-        send_in_retry_.push_back(
-            {task_copy, target_node_id, std::chrono::steady_clock::now()});
-      }
-      continue;
-    }
-
-    // Get or create persistent Lightbeam client using connection pool
-    auto *config_manager = CLIO_CONFIG_MANAGER;
-    int port = static_cast<int>(config_manager->GetPort());
-    ctp::lbm::Transport *lbm_transport =
-        ipc_manager->GetOrCreateClient(target_host->ip_address, port);
-
-    if (!lbm_transport) {
-      HLOG(kError, "[SendIn] Task {} FAILED: Could not get client for {}:{}",
-           origin_task->task_id_, target_host->ip_address, port);
-      ipc_manager->SetDead(target_node_id);
-      {
-        std::lock_guard<std::mutex> _rqlk(retry_queues_mutex_);
-        send_in_retry_.push_back(
-            {task_copy, target_node_id, std::chrono::steady_clock::now()});
-      }
-      continue;
-    }
-
-    // Create SaveTaskArchive with SerializeIn mode and lbm_transport
-    chi::SaveTaskArchive archive(chi::MsgType::kSerializeIn, lbm_transport);
-
-    uint64_t ser_t0 = HrtNs();
-    container->SaveTask(task_copy->method_, archive, task_copy);
-    uint64_t ser_dt = HrtNs() - ser_t0;
-
-    // SYNC send: lightbeam copies bulks into ZMQ during this call. After
-    // Send returns ZMQ holds no reference to task_copy's buffers, so the
-    // copy's lifetime (already pinned by origin_task_rctx->subtasks_[i]
-    // until the response is received) is unaffected by ZMQ's I/O thread.
-    ctp::lbm::LbmContext ctx(ctp::lbm::LBM_SYNC);
-    HLOG(kDebug, "[SendIn] Task {} sending to node {} via lightbeam",
-         origin_task->task_id_, target_node_id);
-    uint64_t lbm_t0 = HrtNs();
-    int rc = lbm_transport->Send(archive, ctx);
-    uint64_t lbm_dt = HrtNs() - lbm_t0;
-    HLOG(kDebug, "[SendIn] Task {} lightbeam Send rc={}", origin_task->task_id_,
-         rc);
-
-    if (rc == 0) {
-      auto &st = NetStats(target_node_id);
-      st.sin_count += 1;
-      st.sin_bytes += origin_task->GetRunCtx()
-                          ? origin_task->GetRunCtx()->predicted_stat_.io_size_
-                          : 0;
-      st.sin_lbm_ns += lbm_dt;
-      st.sin_ser_ns += ser_dt;
-    }
-
-    if (rc != 0) {
-      HLOG(kWarning,
-           "[SendIn] Task {} Lightbeam Send rc={} — re-queueing (no dead-mark)",
-           origin_task->task_id_, rc);
-      if (rc != EAGAIN) {
-        ipc_manager->SetDead(target_node_id);
-      }
-      {
-        std::lock_guard<std::mutex> _rqlk(retry_queues_mutex_);
-        send_in_retry_.push_back(
-            {task_copy, target_node_id, std::chrono::steady_clock::now()});
-      }
-      continue;
-    }
-
-    if (origin_task->method_ == Method::kHeartbeatProbe ||
-        origin_task->method_ == Method::kProbeRequest ||
-        origin_task->method_ == Method::kHeartbeat) {
-      HLOG(kInfo,
-           "[ProbeTrace SendIn] method={} task={} -> node={} sent OK",
-           origin_task->method_, origin_task->task_id_, target_node_id);
-    }
-
-    {
-      static std::atomic<size_t> ctr{0};
-      size_t t = ctr.fetch_add(1, std::memory_order_relaxed) + 1;
-      if ((t & 0xff) == 0) {
-        HLOG(kDebug, "[CountSendIn] sent {} cross-node tasks to peers", t);
-      }
-    }
-  }
+  CLIO_IPC->GetRun2Run()->SendIn(origin_task);
 }
 
 /**
  * Helper function: Send task outputs back to origin node
- * @param origin_task Completed task whose outputs need to be sent back
  */
 void Runtime::SendOut(ctp::ipc::FullPtr<chi::Task> origin_task) {
-  auto *ipc_manager = CLIO_IPC;
-  auto *pool_manager = CLIO_POOL_MANAGER;
-
-  // Task lifetime across the zero-copy ZMQ send is handled by lightbeam
-  // (via LbmContext::on_send_complete set just before Send below). The
-  // transport keeps origin_task's serialization buffer alive via an
-  // atomic refcount tied to each zmq_msg_init_data ref; when every bulk
-  // has flushed (or been released via zmq_msg_close on a failure path)
-  // the callback fires and DelTask runs.  No deferred-delete queue,
-  // no time-based deferral — the same event-driven mechanism that
-  // IpcCpu2CpuZmq::RuntimeSend uses on the client-response path.
-
-  // Validate origin_task
-  if (origin_task.IsNull()) {
-    HLOG(kError, "SendOut: origin_task is null");
-    return;
-  }
-
-  // Get the container associated with the origin_task
-  chi::Container *container =
-      pool_manager->GetStaticContainer(origin_task->pool_id_);
-  if (container == nullptr) {
-    HLOG(kError, "SendOut: container not found for pool_id {}",
-         origin_task->pool_id_);
-    return;
-  }
-
-  // Remove task from recv_map as we're completing it.
-  // Erase is idempotent: SendOut can be called again from send_out_retry_
-  // after IsAlive() or Lightbeam Send fails. The first call erases the
-  // entry; subsequent retries no-op here and still attempt the send.
-  // RecvIn writes recv_map_ from net_recv_worker, so the mutex serialises
-  // the erase against concurrent inserts.
-  size_t recv_key = origin_task->task_id_.net_key_ ^
-                    (static_cast<size_t>(origin_task->task_id_.replica_id_) *
-                     0x9e3779b97f4a7c15ULL);
-  {
-    std::lock_guard<std::mutex> lk(recv_map_mutex_);
-    recv_map_.erase(recv_key);
-  }
-
-  // Get return node from pool_query
-  chi::u64 target_node_id = origin_task->pool_query_.GetReturnNode();
-
-  // Check aliveness before sending output back
-  if (!ipc_manager->IsAlive(target_node_id)) {
-    HLOG(kWarning,
-         "[SendOut] Task {} return node {} is dead, queuing for retry",
-         origin_task->task_id_, target_node_id);
-    {
-      std::lock_guard<std::mutex> _rqlk(retry_queues_mutex_);
-      send_out_retry_.push_back(
-          {origin_task, target_node_id, std::chrono::steady_clock::now()});
-    }
-    return;
-  }
-
-  // Get host information
-  const chi::Host *target_host = ipc_manager->GetHost(target_node_id);
-  if (target_host == nullptr) {
-    HLOG(kError, "[SendOut] Task {} FAILED: Host not found for node_id {}",
-         origin_task->task_id_, target_node_id);
-    return;
-  }
-
-  auto *config_manager = CLIO_CONFIG_MANAGER;
-  int port = static_cast<int>(config_manager->GetPort());
-  ctp::lbm::Transport *lbm_transport =
-      ipc_manager->GetOrCreateClient(target_host->ip_address, port);
-
-  if (lbm_transport == nullptr) {
-    HLOG(kError, "[SendOut] Task {} FAILED: Could not get client for {}:{}",
-         origin_task->task_id_, target_host->ip_address, port);
-    ipc_manager->SetDead(target_node_id);
-    {
-      std::lock_guard<std::mutex> _rqlk(retry_queues_mutex_);
-      send_out_retry_.push_back(
-          {origin_task, target_node_id, std::chrono::steady_clock::now()});
-    }
-    return;
-  }
-
-  chi::SaveTaskArchive archive(chi::MsgType::kSerializeOut, lbm_transport);
-
-  uint64_t sout_ser_t0 = HrtNs();
-  container->SaveTask(origin_task->method_, archive, origin_task);
-  uint64_t sout_ser_dt = HrtNs() - sout_ser_t0;
-
-  // SYNC send: lightbeam copies our bulks into ZMQ during the call and
-  // holds no reference to origin_task after Send returns, so DelTask
-  // runs deterministically below — no on_send_complete, no async
-  // lifetime tracking, no UAF race with ZMQ's I/O thread.
-  ctp::lbm::LbmContext ctx(ctp::lbm::LBM_SYNC);
-  uint64_t sout_lbm_t0 = HrtNs();
-  int rc = lbm_transport->Send(archive, ctx);
-  uint64_t sout_lbm_dt = HrtNs() - sout_lbm_t0;
-
-  if (rc == 0) {
-    auto &st = NetStats(target_node_id);
-    st.sout_count += 1;
-    st.sout_bytes += origin_task->GetRunCtx()
-                         ? origin_task->GetRunCtx()->predicted_stat_.io_size_
-                         : 0;
-    st.sout_lbm_ns += sout_lbm_dt;
-    st.sout_ser_ns += sout_ser_dt;
-  }
-  if (rc != 0) {
-    HLOG(kWarning,
-         "[SendOut] Task {} Lightbeam Send rc={} — re-queueing (no dead-mark)",
-         origin_task->task_id_, rc);
-    if (rc != EAGAIN) {
-      ipc_manager->SetDead(target_node_id);
-    }
-    {
-      std::lock_guard<std::mutex> _rqlk(retry_queues_mutex_);
-      send_out_retry_.push_back(
-          {origin_task, target_node_id, std::chrono::steady_clock::now()});
-    }
-    return;
-  }
-
-  if (origin_task->method_ == Method::kHeartbeatProbe ||
-      origin_task->method_ == Method::kProbeRequest ||
-      origin_task->method_ == Method::kHeartbeat) {
-    HLOG(kInfo,
-         "[ProbeTrace SendOut] method={} task={} -> node={} sent OK",
-         origin_task->method_, origin_task->task_id_, target_node_id);
-  }
-
-  {
-    static std::atomic<size_t> ctr{0};
-    size_t t = ctr.fetch_add(1, std::memory_order_relaxed) + 1;
-    if ((t & 0xff) == 0) {
-      HLOG(kDebug, "[CountSendOut] sent {} cross-node responses back", t);
-    }
-  }
-
-  // SYNC send: Send returned, ZMQ has copied every bulk into its own
-  // buffer, no outstanding reference to origin_task remains anywhere
-  // outside us. Safe to DelTask synchronously.
-  container->DelTask(origin_task->method_, origin_task);
+  CLIO_IPC->GetRun2Run()->SendOut(origin_task);
 }
 
 /**
@@ -975,8 +410,8 @@ chi::TaskResume Runtime::Send(ctp::ipc::FullPtr<SendTask> task,
   bool did_send = false;
 
   // Per-tick maintenance: retries and dead-node fanout.
-  ProcessRetryQueues();
-  ScanSendMapTimeouts();
+  CLIO_IPC->GetRun2Run()->ProcessRetryQueues();
+  CLIO_IPC->GetRun2Run()->ScanSendMapTimeouts();
 
   // Snapshot the depth of each priority at function entry so a hot
   // producer can't monopolise this tick.
@@ -1057,7 +492,8 @@ chi::TaskResume Runtime::Send(ctp::ipc::FullPtr<SendTask> task,
   rctx.did_work_ = did_send;
   {
     auto *cfg_ipc = CLIO_IPC;
-    DumpNetStatsIfDue(cfg_ipc->GetNodeId(), cfg_ipc->GetNumHosts());
+    CLIO_IPC->GetRun2Run()->DumpNetStats(cfg_ipc->GetNodeId(),
+                                          cfg_ipc->GetNumHosts());
   }
   task->SetReturnCode(0);
   CLIO_CO_RETURN;
@@ -1066,361 +502,21 @@ chi::TaskResume Runtime::Send(ctp::ipc::FullPtr<SendTask> task,
 
 /**
  * Helper function: Receive task inputs from remote node
- * @param task RecvTask containing control information
- * @param archive Already-parsed LoadTaskArchive containing task info
- * @param lbm_transport Lightbeam server for receiving bulk data
  */
 void Runtime::RecvIn(ctp::ipc::FullPtr<RecvTask> task,
                      chi::LoadTaskArchive &archive,
                      ctp::lbm::Transport *lbm_transport) {
-  auto *ipc_manager = CLIO_IPC;
-  auto *pool_manager = CLIO_POOL_MANAGER;
-
-  const auto &task_infos = archive.GetTaskInfos();
-
-  // If no tasks to receive
-  if (task_infos.empty()) {
-    task->SetReturnCode(0);
-    return;
-  }
-
-  for (size_t task_idx = 0; task_idx < task_infos.size(); ++task_idx) {
-    const auto &task_info = task_infos[task_idx];
-
-    // Get container associated with PoolId
-    chi::Container *container =
-        pool_manager->GetStaticContainer(task_info.pool_id_);
-    if (!container) {
-      HLOG(kError, "Admin: Container not found for pool_id {}",
-           task_info.pool_id_);
-      continue;
-    }
-
-    // Instrumentation: time the deserialize+alloc step so the
-    // dump can show whether the recv-side cost is dominated by
-    // bulk copy or by task object construction.
-    uint64_t des_t0 = HrtNs();
-    // Call AllocLoadTask to allocate and deserialize the task
-    ctp::ipc::FullPtr<chi::Task> task_ptr =
-        container->AllocLoadTask(task_info.method_id_, archive);
-    uint64_t des_dt = HrtNs() - des_t0;
-
-    if (task_ptr.IsNull()) {
-      HLOG(kError, "Admin: Failed to load task");
-      continue;
-    }
-    {
-      chi::u64 from_node = task_ptr->pool_query_.GetReturnNode();
-      auto &st = NetStats(from_node);
-      st.rin_count += 1;
-      st.rin_bytes += task_ptr->GetRunCtx()
-                          ? task_ptr->GetRunCtx()->predicted_stat_.io_size_
-                          : 0;
-      st.rin_des_ns += des_dt;
-    }
-
-    // If the sender is a node we marked kDead, mark it alive.
-    // This handles restarted nodes: their SWIM probes reach us here,
-    // letting us rediscover them without requiring --induct.
-    // Only trigger on kDead (not kProbeFailed/kSuspected) to avoid
-    // flip-flopping during normal SWIM failure detection.
-    chi::u64 sender_node = task_ptr->pool_query_.GetReturnNode();
-    if (sender_node != ipc_manager->GetNodeId() &&
-        ipc_manager->GetNodeState(sender_node) == chi::NodeState::kDead) {
-      HLOG(kInfo, "[RecvIn] Received task from dead node {}, marking alive",
-           sender_node);
-      // Flush stale retry entries BEFORE marking alive so
-      // ProcessRetryQueues doesn't resend old tasks to the fresh runtime.
-      FlushStaleStateForNode(sender_node);
-      ipc_manager->SetAlive(sender_node);
-    }
-
-    // Mark task as remote. TASK_DATA_OWNER is set only if the archive
-    // had to AllocateBuffer for any BULK_EXPOSE bulk (typical read
-    // response path: GetBlob's blob_data_ is allocated daemon-side). If
-    // the input was BULK_XFER (write data shipped from peer), the bulk
-    // buffer lives in ZMQ-owned memory and FreeBuffer on it would crash,
-    // so leave TASK_DATA_OWNER clear in that case.
-    //
-    // TASK_RUN_CTX_EXISTS and TASK_STARTED must be cleared so the
-    // receiving worker allocates a fresh RunContext via BeginTask.
-    chi::u32 set_flags = TASK_REMOTE;
-    if (archive.daemon_allocated_bulk_count_ > 0) {
-      set_flags |= TASK_DATA_OWNER;
-    }
-    task_ptr->SetFlags(set_flags);
-    task_ptr->ClearFlags(TASK_PERIODIC | TASK_ROUTED |
-                         TASK_RUN_CTX_EXISTS | TASK_STARTED);
-
-    // Add task to recv_map for later lookup
-    // Key combines net_key and replica_id so multiple replicas targeting the
-    // same node (e.g., after container migration) get distinct entries.
-    // RecvIn runs on net_recv_worker; SendOut reads/erases recv_map_ from
-    // net_send_worker.
-    size_t recv_key = task_ptr->task_id_.net_key_ ^
-                      (static_cast<size_t>(task_ptr->task_id_.replica_id_) *
-                       0x9e3779b97f4a7c15ULL);
-    {
-      std::lock_guard<std::mutex> lk(recv_map_mutex_);
-      recv_map_[recv_key] = task_ptr;
-    }
-
-    HLOG(kDebug, "[RecvIn] Task {} method={} pool_id={} dispatching to workers",
-         task_ptr->task_id_, task_ptr->method_, task_ptr->pool_id_);
-
-    if (task_ptr->method_ == Method::kHeartbeatProbe ||
-        task_ptr->method_ == Method::kProbeRequest ||
-        task_ptr->method_ == Method::kHeartbeat) {
-      HLOG(kInfo,
-           "[ProbeTrace RecvIn] method={} task={} from node={}",
-           task_ptr->method_, task_ptr->task_id_, sender_node);
-    }
-
-    // Dispatch the received task directly onto a worker lane.
-    //
-    // We deliberately do NOT call ipc_manager->Send here because RecvIn now
-    // runs on a dedicated std::thread (CLIO_CUR_WORKER == nullptr). Send's
-    // non-worker branch routes through IpcManager::RouteTask, which respects
-    // pool_query_ and may re-broadcast a task that arrived at us because
-    // we cleared TASK_ROUTED above. Mirror IpcCpu2Self::ClientSend's worker
-    // branch directly: build a pointer-Future, pick a lane via
-    // scheduler->ClientMapTask, push, and AwakenWorker if the lane was empty.
-    // No parent RunContext is set (there is no current run context on a
-    // recv thread, and these are top-level remote tasks anyway).
-    chi::Future<chi::Task> future =
-        ipc_manager->MakePointerFuture(task_ptr);
-    if (future.GetFutureShm().IsNull()) {
-      HLOG(kError, "[RecvIn] MakePointerFuture failed for task {}",
-           task_ptr->task_id_);
-      continue;
-    }
-    if (!task_ptr->task_flags_.Any(TASK_RUN_CTX_EXISTS)) {
-      ipc_manager->BeginTask(future, nullptr, nullptr);
-    }
-    // ROUTED so the worker doesn't re-route this task to peers when it
-    // sees pool_query_ pointing at another node.
-    task_ptr->SetFlags(TASK_ROUTED);
-
-    if (ipc_manager->GetScheduler() != nullptr) {
-      chi::u32 lane_id =
-          ipc_manager->GetScheduler()->ClientMapTask(ipc_manager, future);
-      auto *worker_queues = ipc_manager->GetTaskQueue();
-      if (worker_queues) {
-        auto &dest_lane = worker_queues->GetLane(lane_id, 0);
-        dest_lane.Push(future);
-        // Always signal — see ipc_cpu2cpu_impl.h for the race.
-        ipc_manager->AwakenWorker(&dest_lane);
-      }
-    }
-  }
-
+  CLIO_IPC->GetRun2Run()->RecvIn(archive, lbm_transport);
   task->SetReturnCode(0);
 }
 
 /**
  * Helper function: Receive task outputs from remote node
- * @param task RecvTask containing control information
- * @param archive Already-parsed LoadTaskArchive containing task info
- * @param lbm_transport Lightbeam server for receiving bulk data
  */
 void Runtime::RecvOut(ctp::ipc::FullPtr<RecvTask> task,
                       chi::LoadTaskArchive &archive,
                       ctp::lbm::Transport *lbm_transport) {
-  auto *pool_manager = CLIO_POOL_MANAGER;
-
-  const auto &task_infos = archive.GetTaskInfos();
-
-  // If no task outputs to receive
-  if (task_infos.empty()) {
-    task->SetReturnCode(0);
-    return;
-  }
-
-  // Set lbm_transport in archive for bulk transfer exposure in output mode
-  archive.SetTransport(lbm_transport);
-
-  // First pass: Deserialize to expose buffers
-  // LoadTask will call ar.bulk() which will expose the pointers and populate
-  // archive.recv
-  for (size_t task_idx = 0; task_idx < task_infos.size(); ++task_idx) {
-    const auto &task_info = task_infos[task_idx];
-
-    // Locate origin task from send_map using net_key
-    size_t net_key = task_info.task_id_.net_key_;
-
-    // SendIn (net_send_worker) populates send_map_; copy the value out
-    // under the lock so the rest of the work happens without holding it.
-    ctp::ipc::FullPtr<chi::Task> origin_task;
-    {
-      std::lock_guard<std::mutex> lk(send_map_mutex_);
-      auto send_it = send_map_.find(net_key);
-      if (send_it == nullptr) {
-        HLOG(kError,
-             "[RecvOut] Task {} FAILED: Origin task not found in send_map "
-             "(size: {}) with net_key {}",
-             task_info.task_id_, send_map_.size(), net_key);
-        task->SetReturnCode(5);
-        return;
-      }
-      origin_task = *send_it;
-    }
-    if (!origin_task->GetRunCtx()) {
-      HLOG(kError, "Admin: origin_task has no RunContext");
-      task->SetReturnCode(6);
-      return;
-    }
-    chi::RunContext *origin_rctx = origin_task->GetRunCtx();
-
-    // Locate replica in origin's run_ctx using replica_id
-    chi::u32 replica_id = task_info.task_id_.replica_id_;
-    if (replica_id >= origin_rctx->subtasks_.size()) {
-      HLOG(kError, "Admin: Invalid replica_id {} (subtasks size: {})",
-           replica_id, origin_rctx->subtasks_.size());
-      task->SetReturnCode(7);
-      return;
-    }
-
-    ctp::ipc::FullPtr<chi::Task> replica = origin_rctx->subtasks_[replica_id];
-
-    // Get the container associated with the origin task
-    chi::Container *container =
-        pool_manager->GetStaticContainer(origin_task->pool_id_);
-    if (!container) {
-      HLOG(kError, "Admin: Container not found for pool_id {}",
-           origin_task->pool_id_);
-      task->SetReturnCode(8);
-      return;
-    }
-
-    // Deserialize outputs directly into the replica task using LoadTask
-    // This exposes buffers via ar.bulk() and populates archive.recv.
-    // Instrumentation: time the deserialize for rout accounting.
-    uint64_t rout_des_t0 = HrtNs();
-    container->LoadTask(origin_task->method_, archive, replica);
-    uint64_t rout_des_dt = HrtNs() - rout_des_t0;
-    {
-      // task_info has sender node via task_id_.node_id_ (the origin node).
-      chi::u64 from_node = task_info.task_id_.node_id_;
-      auto &st = NetStats(from_node);
-      st.rout_count += 1;
-      st.rout_bytes += origin_task->GetRunCtx()
-                           ? origin_task->GetRunCtx()->predicted_stat_.io_size_
-                           : 0;
-      st.rout_des_ns += rout_des_dt;
-    }
-
-  }
-
-  // Second pass: Aggregate results
-  for (size_t task_idx = 0; task_idx < task_infos.size(); ++task_idx) {
-    const auto &task_info = task_infos[task_idx];
-
-    // Locate origin task from send_map using net_key. Same locking note as
-    // the first pass — copy out under the lock, then process unlocked.
-    size_t net_key = task_info.task_id_.net_key_;
-    ctp::ipc::FullPtr<chi::Task> origin_task;
-    {
-      std::lock_guard<std::mutex> lk(send_map_mutex_);
-      auto send_it = send_map_.find(net_key);
-      if (send_it == nullptr) {
-        HLOG(kError, "Admin: Origin task not found in send_map with net_key {}",
-             net_key);
-        continue;
-      }
-      origin_task = *send_it;
-    }
-    if (!origin_task->GetRunCtx()) {
-      HLOG(kError, "Admin: origin_task has no RunContext");
-      continue;
-    }
-    chi::RunContext *origin_rctx = origin_task->GetRunCtx();
-
-    // Locate replica in origin's run_ctx using replica_id
-    chi::u32 replica_id = task_info.task_id_.replica_id_;
-    if (replica_id >= origin_rctx->subtasks_.size()) {
-      HLOG(kError, "Admin: Invalid replica_id {} (subtasks size: {})",
-           replica_id, origin_rctx->subtasks_.size());
-      continue;
-    }
-
-    ctp::ipc::FullPtr<chi::Task> replica = origin_rctx->subtasks_[replica_id];
-
-    // Get the container associated with the origin task
-    chi::Container *container =
-        pool_manager->GetStaticContainer(origin_task->pool_id_);
-    if (!container) {
-      HLOG(kError, "Admin: Container not found for pool_id {}",
-           origin_task->pool_id_);
-      continue;
-    }
-
-    // Aggregate replica results into origin task via container dispatch
-    container->Aggregate(origin_task->method_, origin_task, replica);
-
-    HLOG(kDebug, "[RecvOut] Task {}", origin_task->task_id_);
-
-    // Increment completed replicas counter in origin's rctx.
-    // Use the fetch_add return so we observe THIS thread's increment —
-    // a separate load can race a parallel RecvOut/SendIn-skip increment
-    // and miss the "all done" trigger condition.
-    chi::u32 completed = origin_rctx->completed_replicas_.fetch_add(1) + 1;
-
-    // If all replicas completed
-    if (completed == origin_rctx->subtasks_.size()) {
-      // Get pool manager to access container
-      auto *pool_manager = CLIO_POOL_MANAGER;
-      chi::Container *container =
-          pool_manager->GetStaticContainer(origin_task->pool_id_);
-
-      // Unmark TASK_DATA_OWNER before deleting replicas to avoid freeing the
-      // same data pointers twice. Delete all origin_task replicas using
-      // container->DelTask() to avoid memory leak
-      for (const auto &origin_task_ptr : origin_rctx->subtasks_) {
-        origin_task_ptr->ClearFlags(TASK_DATA_OWNER);
-        container->DelTask(origin_task->method_, origin_task_ptr);
-      }
-
-      // Clear subtasks vector after deleting tasks
-      origin_rctx->subtasks_.clear();
-
-      // Remove origin from send_map; serialised against SendIn insertions
-      // and ScanSendMapTimeouts iteration on the send worker.
-      {
-        std::lock_guard<std::mutex> lk(send_map_mutex_);
-        send_map_.erase(net_key);
-      }
-
-      // Set container in origin RunContext (may be null if task was routed
-      // globally without passing through RouteLocal — e.g. when the runtime
-      // is started with CLIO_FORCE_NET=1)
-      if (container) {
-        origin_rctx->container_ = container;
-      }
-
-      // Complete the origin task via EndTask. When RecvOut runs on the
-      // dedicated peer recv thread (CLIO_CUR_WORKER == nullptr), borrow the
-      // net_recv_worker pointer for the call. EndTask's per-worker state
-      // updates (num_tasks_processed_, load_) are stats only — the
-      // correctness-bearing paths (EnqueueNetTask for remote origin,
-      // IpcCpu2Self::RuntimeSend for local origin) are thread-safe and
-      // don't depend on CLIO_CUR_WORKER being correct. The net_recv_worker
-      // already serialized incoming network responses pre-split, so this
-      // approximates the prior accounting.
-      auto *worker = CLIO_CUR_WORKER;
-      if (worker == nullptr) {
-        auto *scheduler = CLIO_IPC->GetScheduler();
-        worker = scheduler ? scheduler->GetNetRecvWorker() : nullptr;
-      }
-      if (worker) {
-        worker->EndTask(origin_task, origin_rctx, true);
-      } else {
-        HLOG(kError,
-             "[RecvOut] No worker available to call EndTask for task {}",
-             origin_task->task_id_);
-      }
-    }
-  }
-
+  CLIO_IPC->GetRun2Run()->RecvOut(archive, lbm_transport);
   task->SetReturnCode(0);
 }
 
@@ -2392,279 +1488,6 @@ chi::TaskResume Runtime::MigrateContainers(
   CLIO_TASK_BODY_END
 }
 
-/**
- * Attempt to send a retried task to the given node
- * @param entry The retry entry containing the task
- * @param node_id The node to send to
- * @return true if send succeeded
- */
-bool Runtime::RetrySendToNode(RetryEntry &entry, chi::u64 node_id) {
-  auto *ipc_manager = CLIO_IPC;
-  auto *pool_manager = CLIO_POOL_MANAGER;
-  auto *config_manager = CLIO_CONFIG_MANAGER;
-
-  const chi::Host *target_host = ipc_manager->GetHost(node_id);
-  if (!target_host) {
-    return false;
-  }
-  int port = static_cast<int>(config_manager->GetPort());
-  ctp::lbm::Transport *lbm_transport =
-      ipc_manager->GetOrCreateClient(target_host->ip_address, port);
-  if (!lbm_transport) {
-    return false;
-  }
-  chi::Container *container =
-      pool_manager->GetStaticContainer(entry.task->pool_id_);
-  if (!container) {
-    return false;
-  }
-  chi::SaveTaskArchive archive(chi::MsgType::kSerializeIn, lbm_transport);
-  container->SaveTask(entry.task->method_, archive, entry.task);
-  ctp::lbm::LbmContext ctx(0);
-  int rc = lbm_transport->Send(archive, ctx);
-  return rc == 0;
-}
-
-/**
- * Re-resolve target node for a retried task whose original target is dead.
- * Uses the task's pool_query_ to look up the current container→node mapping
- * from the address_map_, which may have been updated by recovery.
- * @param entry The retry entry to re-resolve
- * @return New node ID, or 0 if re-resolution failed
- */
-chi::u64 Runtime::RerouteRetryEntry(RetryEntry &entry) {
-  auto *pool_manager = CLIO_POOL_MANAGER;
-  const chi::PoolQuery &query = entry.task->pool_query_;
-
-  if (query.IsDirectIdMode()) {
-    chi::ContainerId container_id = query.GetContainerId();
-    chi::u32 new_node =
-        pool_manager->GetContainerNodeId(entry.task->pool_id_, container_id);
-    if (new_node != 0 && new_node != entry.target_node_id) {
-      return new_node;
-    }
-  } else if (query.IsRangeMode()) {
-    chi::u32 offset = query.GetRangeOffset();
-    chi::ContainerId container_id(offset);
-    chi::u32 new_node =
-        pool_manager->GetContainerNodeId(entry.task->pool_id_, container_id);
-    if (new_node != 0 && new_node != entry.target_node_id) {
-      return new_node;
-    }
-  }
-  return 0;
-}
-
-void Runtime::ProcessRetryQueues() {
-  auto *ipc_manager = CLIO_IPC;
-  auto now = std::chrono::steady_clock::now();
-
-  // Sender threads push to send_*_retry_ on EAGAIN / failure.
-  // ProcessRetryQueues runs on the SendPoll periodic thread; serialise
-  // iteration against concurrent pushes. unique_lock so we can release
-  // around the recursive SendOut() call (SendOut may want this lock).
-  std::unique_lock<std::mutex> _rqlk(retry_queues_mutex_);
-
-  // Process send_in retry queue
-  auto it = send_in_retry_.begin();
-  while (it != send_in_retry_.end()) {
-    float elapsed = std::chrono::duration<float>(now - it->enqueued_at).count();
-    float task_timeout = kRetryTimeoutSec;
-    float task_net_timeout = it->task->pool_query_.GetNetTimeout();
-    if (task_net_timeout >= 0) {
-      task_timeout = task_net_timeout;
-    }
-    if (elapsed >= task_timeout) {
-      // Timeout: mark task as failed
-      HLOG(kError, "[RetryQueue] SendIn task timed out after {}s for node {}",
-           elapsed, it->target_node_id);
-      it->task->SetReturnCode(kNetworkTimeoutRC);
-      it = send_in_retry_.erase(it);
-    } else if (ipc_manager->IsAlive(it->target_node_id)) {
-      // Original node came back: retry the send
-      if (RetrySendToNode(*it, it->target_node_id)) {
-        HLOG(kInfo, "[RetryQueue] SendIn retry succeeded for node {}",
-             it->target_node_id);
-        it = send_in_retry_.erase(it);
-        continue;
-      }
-      // Retry failed, keep in queue
-      ++it;
-    } else {
-      // Original node still dead — try re-routing via updated address_map
-      chi::u64 new_node = RerouteRetryEntry(*it);
-      if (new_node != 0 && ipc_manager->IsAlive(new_node)) {
-        HLOG(kInfo,
-             "[RetryQueue] Re-routing task from dead node {} to "
-             "recovered node {}",
-             it->target_node_id, new_node);
-        it->target_node_id = new_node;
-        if (RetrySendToNode(*it, new_node)) {
-          HLOG(kInfo,
-               "[RetryQueue] SendIn re-routed retry succeeded "
-               "for node {}",
-               new_node);
-          it = send_in_retry_.erase(it);
-          continue;
-        }
-      }
-      ++it;
-    }
-  }
-
-  // Process send_out retry queue
-  it = send_out_retry_.begin();
-  while (it != send_out_retry_.end()) {
-    float elapsed = std::chrono::duration<float>(now - it->enqueued_at).count();
-    float out_task_timeout = kRetryTimeoutSec;
-    float out_task_net_timeout = it->task->pool_query_.GetNetTimeout();
-    if (out_task_net_timeout >= 0) {
-      out_task_timeout = out_task_net_timeout;
-    }
-    if (elapsed >= out_task_timeout) {
-      HLOG(kError, "[RetryQueue] SendOut task timed out after {}s for node {}",
-           elapsed, it->target_node_id);
-      // For send_out, the result is lost; origin will timeout
-      it = send_out_retry_.erase(it);
-    } else if (ipc_manager->IsAlive(it->target_node_id)) {
-      // Node came back: retry by calling SendOut. SendOut may push to
-      // send_out_retry_ on a fresh failure — release the lock around it
-      // to avoid self-deadlock, then restart iteration from begin().
-      ctp::ipc::FullPtr<chi::Task> retry_task = it->task;
-      it = send_out_retry_.erase(it);
-      _rqlk.unlock();
-      SendOut(retry_task);
-      _rqlk.lock();
-      it = send_out_retry_.begin();
-      continue;
-      ++it;
-    }
-  }
-}
-
-void Runtime::ScanSendMapTimeouts() {
-  auto *ipc_manager = CLIO_IPC;
-  auto now = std::chrono::steady_clock::now();
-
-  // Iterate dead nodes and check if any send_map_ entries target them
-  const auto &dead_nodes = ipc_manager->GetDeadNodes();
-  if (dead_nodes.empty()) return;
-
-  // Build map of dead node IDs to their detected_at times for per-task checks
-  std::unordered_map<chi::u64, std::chrono::steady_clock::time_point> dead_map;
-  for (const auto &entry : dead_nodes) {
-    dead_map[entry.node_id] = entry.detected_at;
-  }
-
-  // Scan send_map_ for tasks targeting dead nodes using for_each.
-  //
-  // ScanSendMapTimeouts runs on net_send_worker (same thread as SendIn
-  // inserts), but RecvOut on net_recv_worker erases entries concurrently.
-  // Hold the lock for the iteration, but EndTask is called only on
-  // collected keys and may do nontrivial work, so the actual EndTask +
-  // erase pass runs outside the iteration scope. This keeps the critical
-  // section bounded to a hash-bucket walk.
-  std::vector<std::pair<size_t, ctp::ipc::FullPtr<chi::Task>>> to_complete;
-  {
-    std::lock_guard<std::mutex> lk(send_map_mutex_);
-    send_map_.for_each(
-        [&](const size_t &key, ctp::ipc::FullPtr<chi::Task> &origin_task) {
-          if (origin_task.IsNull() || !origin_task->GetRunCtx()) return;
-
-          chi::RunContext *rctx = origin_task->GetRunCtx();
-          // Use per-task timeout if set, otherwise kRetryTimeoutSec
-          float task_timeout = kRetryTimeoutSec;
-          float task_net_timeout = origin_task->pool_query_.GetNetTimeout();
-          if (task_net_timeout >= 0) {
-            task_timeout = task_net_timeout;
-          }
-
-          // Check if any replica targets a dead node that has exceeded the
-          // timeout
-          bool any_timed_out = false;
-          for (const auto &pq : rctx->pool_queries_) {
-            if (pq.IsPhysicalMode()) {
-              auto dit = dead_map.find(pq.GetNodeId());
-              if (dit != dead_map.end()) {
-                float dead_elapsed =
-                    std::chrono::duration<float>(now - dit->second).count();
-                if (dead_elapsed >= task_timeout) {
-                  any_timed_out = true;
-                  break;
-                }
-              }
-            }
-          }
-
-          if (any_timed_out) {
-            to_complete.emplace_back(key, origin_task);
-          }
-        });
-  }
-
-  // Finalise timed-out tasks outside the map lock to keep EndTask from
-  // executing under it (EndTask can re-enter scheduling code).
-  for (auto &entry : to_complete) {
-    auto &origin_task = entry.second;
-    chi::RunContext *rctx = origin_task->GetRunCtx();
-    if (!rctx) continue;
-    HLOG(kError,
-         "[ScanSendMapTimeouts] Task {} timed out waiting for dead node",
-         origin_task->task_id_);
-    origin_task->SetReturnCode(kNetworkTimeoutRC);
-    auto *worker = CLIO_CUR_WORKER;
-    worker->EndTask(origin_task, rctx, true);
-  }
-
-  if (!to_complete.empty()) {
-    std::lock_guard<std::mutex> lk(send_map_mutex_);
-    for (auto &entry : to_complete) {
-      send_map_.erase(entry.first);
-    }
-  }
-}
-
-void Runtime::FlushStaleStateForNode(chi::u64 node_id) {
-  std::lock_guard<std::mutex> _rqlk(retry_queues_mutex_);
-  // 1. Discard send_in retry entries targeting this node.
-  //    For each discarded entry, increment the origin task's
-  //    completed_replicas so broadcast origins can still complete.
-  for (auto it = send_in_retry_.begin(); it != send_in_retry_.end();) {
-    if (it->target_node_id == node_id) {
-      size_t net_key = it->task->task_id_.net_key_;
-      ctp::ipc::FullPtr<chi::Task> origin;
-      {
-        std::lock_guard<std::mutex> lk(send_map_mutex_);
-        auto send_it = send_map_.find(net_key);
-        if (send_it != nullptr) {
-          origin = *send_it;
-        }
-      }
-      if (!origin.IsNull() && origin->GetRunCtx()) {
-        origin->GetRunCtx()->completed_replicas_++;
-      }
-      HLOG(kInfo,
-           "[FlushStale] Discarding SendIn retry for restarted node {}",
-           node_id);
-      it = send_in_retry_.erase(it);
-    } else {
-      ++it;
-    }
-  }
-
-  // 2. Discard send_out retry entries targeting this node.
-  //    These are responses destined for the old incarnation; drop them.
-  for (auto it = send_out_retry_.begin(); it != send_out_retry_.end();) {
-    if (it->target_node_id == node_id) {
-      HLOG(kInfo,
-           "[FlushStale] Discarding SendOut retry for restarted node {}",
-           node_id);
-      it = send_out_retry_.erase(it);
-    } else {
-      ++it;
-    }
-  }
-}
 
 chi::TaskResume Runtime::Heartbeat(ctp::ipc::FullPtr<HeartbeatTask> task,
                                    chi::RunContext &rctx) {
@@ -2982,18 +1805,8 @@ chi::TaskStat Runtime::GetTaskStats(const chi::Task *task) const {
 }
 
 chi::u64 Runtime::GetWorkRemaining() const {
-  // Accessed from non-net workers (the scheduler/shutdown path); briefly
-  // acquire both map locks for a consistent snapshot.
-  size_t send_size, recv_size;
-  {
-    std::lock_guard<std::mutex> lk(send_map_mutex_);
-    send_size = send_map_.size();
-  }
-  {
-    std::lock_guard<std::mutex> lk(recv_map_mutex_);
-    recv_size = recv_map_.size();
-  }
-  return send_size + recv_size;
+  auto *run2run = CLIO_IPC->GetRun2Run();
+  return run2run->GetSendMapSize() + run2run->GetRecvMapSize();
 }
 
 //===========================================================================
